+++ tests/kube/ingress.test.ts
@@ -181,6 +181,55 @@ describe('kubernetes ingresses', () => {
       await bindings.data(resource);
       expect(Icinga.applyHost.mock.calls.length).toBe(0);  
     });
+
+    it('skip resource with invalid kind', async () => {
+      let instance = new Ingress(Logger, Node, Icinga);
+
+      fixture.kind = 'foo';
+      var resource = {  
+        type: 'ADDED', 
+        object: fixture
+      };
+
+      var bindings = {};
+      var json = {
+        on: function(name, callback) {
+          bindings[name] = callback.bind(instance);
+        }
+      };
+    
+      await instance.kubeListener(() => {
+        return json;
+      });
+
+      var result = await bindings.data(resource);
+      expect(result).toBe(false);  
+    });
+    
+    it('skip resource kube-icinga/discover===false', async () => {
+      let instance = new Ingress(Logger, Node, Icinga);
+
+      fixture.metadata.annotations['kube-icinga/discover'] = 'false';
+      var resource = {  
+        type: 'ADDED', 
+        object: fixture
+      };
+
+      var bindings = {};
+      var json = {
+        on: function(name, callback) {
+          bindings[name] = callback.bind(instance);
+        }
+      };
+    
+      await instance.kubeListener(() => {
+        return json;
+      });
+
+      var result = await bindings.data(resource);
+      expect(result).toBe(false);  
+      });
+    });      
   });
   
   describe('add ingress object with dummy host', () => {