+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 **Date**: Tue Mar 19 2019 16:01:20 CET 2019
 
 * [FEATURE] Change servicegroup definition #17
-* Enable provisioning on a resource basis #14
+* [FEATURE] Enable provisioning on a resource basis #14
+* [CHANGE] Added more unit tests
+* [CHANGE] Headless services are ignored
 
 
 ## 2.0.1